% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosPlot2.R
\name{Circos_plot_2}
\alias{Circos_plot_2}
\alias{CircosPlot2}
\title{Circos plot 2}
\usage{
CircosPlot2(
  LRscorematrix,
  LRscore_conditions,
  L_cluster,
  R_cluster,
  LR_cluster,
  WhichLRscore,
  color_L,
  color_R,
  color_LR,
  width_same_ligand_cluster,
  width_different_ligand_cluster,
  width_ligand_receptor,
  width_same_receptor_cluster,
  width_different_receptor_cluster,
  cplotthresh,
  cex
)
}
\arguments{
\item{LRscorematrix}{A matrix with columns "ligand", "receptor" and the condition names where the condition columns contain the LRscores of each ligand-receptor pairs in each condition.}

\item{LRscore_conditions}{A vector of conditions to be used when calculating the maximum or mean LR expr_score for each LR pair acorss conditions}

\item{L_cluster}{The predefined vectors of L, R and LRpair clusters.}

\item{R_cluster}{The predefined vectors of L, R and LRpair clusters.}

\item{LR_cluster}{The predefined vectors of L, R and LRpair clusters.}

\item{WhichLRscore}{The LR expr_score to be used in the plot, which is 'LRscore_max' (for each LR pair, the maximum LR expr_score across LRscore_conditions) or 'LRscore_mean' (for each LR pair, the average LR expr_score across LRscore_conditions).}

\item{color_L}{A vector of colors of the ligand clusters, with the name of each element the name of the corresponding cluster}

\item{color_R}{A vector of colors of the receptor clusters, with the name of each element the name of the corresponding cluster}

\item{color_LR}{A vector of colors of the LR clusters, with the name of each element the name of the corresponding cluster}

\item{width_same_ligand_cluster}{Gap width between ligand nodes in the same cluster}

\item{width_different_ligand_cluster}{Gap width between different ligand clusters}

\item{width_ligand_receptor}{Gap width between the ligand nodes and the receptor nodes}

\item{width_same_receptor_cluster}{Gap width between receptor nodes in the same cluster}

\item{width_different_receptor_cluster}{Gap width between different receptor clusters}

\item{cplotthresh}{A number. Only plot L-R edges in the circos plot with LRscore > cplotthresh}

\item{cex}{Font size}
}
\description{
Circos plot 2
}
