% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_receptor_target_matrix.R
\name{construct_receptor_target_matrix}
\alias{construct_receptor_target_matrix}
\title{Construct a matrix containing receptor-target regulatory probability scores}
\usage{
construct_receptor_target_matrix(
  weighted_networks,
  receptors,
  rtf_cutoff = 0.99,
  damping_factor = 0.5,
  secondary_targets = FALSE,
  receptors_as_cols = TRUE,
  remove_direct_links = "no"
)
}
\arguments{
\item{weighted_networks}{A list of two weighted networks sig and gr (signaling network and gene regulatory network), in data frame/tibble with columns "from", "to" and "weight".}

\item{receptors}{A list of receptor gene symbols.}

\item{rtf_cutoff}{Receptor-tf scores lower than the "rtf_cutoff" quantile will be set to 0.}

\item{damping_factor}{A number between 0 and 1, the probability that the random walker in the PPR algorithm will continue the walk on the graph.}

\item{secondary_targets}{"TRUE" or "FALSE", indicates whether putative secondary targets should be included.}

\item{receptors_as_cols}{"TRUE" or "FALSE", indicates whether receptors should be in columns of the matrix and targets in rows or vice versa.}

\item{remove_direct_links}{Indicates whether direct ligand-target and receptor-target links in the gene regulatory network should be kept or not. "no": keep links; "ligand": remove direct ligand-target links; "ligand-receptor": remove both direct ligand-target and receptor-target links.}
}
\value{
a matrix containing receptor-target scores
}
\description{
Construct a matrix containing receptor-target regulatory probability scores
}
