% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_receptor_tf_matrix.R
\name{construct_receptor_tf_matrix}
\alias{construct_receptor_tf_matrix}
\title{Construct a matrix containing receptor-tf probability scores}
\usage{
construct_receptor_tf_matrix(
  weighted_networks,
  receptors,
  rtf_cutoff = 0.99,
  damping_factor = 0.5,
  receptors_as_cols = FALSE
)
}
\arguments{
\item{weighted_networks}{A list of two weighted networks sig and gr (signaling network and gene regulatory network), in data frame/tibble with columns "from", "to" and "weight".}

\item{receptors}{A list of receptor gene symbols.}

\item{rtf_cutoff}{Receptor-tf scores lower than the "rtf_cutoff" quantile will be set to 0.}

\item{damping_factor}{A number between 0 and 1, the probability that the random walker in the PPR algorithm will continue the walk on the graph.}

\item{receptors_as_cols}{"TRUE" or "FALSE", indicates whether receptors should be in columns of the matrix and targets in rows or vice versa.}
}
\value{
A matrix containing receptor-tf scores
}
\description{
Construct a matrix containing receptor-tf probability scores
}
