% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_lr.R
\name{get_expr_lr}
\alias{get_expr_lr}
\title{Identify "expressed" ligand-receptor pairs}
\usage{
get_expr_lr(lr_network, thresh_expr_from, thresh_expr_to, conditions)
}
\arguments{
\item{lr_network}{A ligand-receptor network in matrix with two columns "from" and "to".}

\item{thresh_expr_from}{A matrix of 0s and 1s, indicates whether each gene is "expressed" in each condition in the sender cells of this step}

\item{thresh_expr_to}{A matrix of 0s and 1s, indicates whether each gene is "expressed" in each condition in the receiver cells of this step}

\item{conditions}{A vector of conditions of interest}
}
\value{
lr_expr: A list with two elements:
 lr_expr$eachcondition: A list, each element is a ligand-receptor network matrix with two columns "from" and "to" containing "expressed" ligand-receptor pairs in one condition
 lr_expr$bind: A ligand-receptor matrix with two columns "from" and "to" containing "expressed" ligand-receptor pairs in at least one condition
}
\description{
Identify "expressed" ligand-receptor pairs
}
