% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LRL.R
\name{get_LRL}
\alias{get_LRL}
\title{Get the LRloop L1_R1-L2_R2 pairs}
\usage{
get_LRL(
  lr_expr_ct2_to_ct1,
  lr_expr_ct1_to_ct2,
  ligand_target_matrix_binary_ct2_to_ct1,
  ligand_target_matrix_binary_ct1_to_ct2,
  receptor_target_matrix_binary_ct2_to_ct1,
  receptor_target_matrix_binary_ct1_to_ct2
)
}
\arguments{
\item{lr_expr_ct2_to_ct1}{A ligand-receptor network list resulted from the function "get_expr_lr", signaling from ct2 to ct1}

\item{lr_expr_ct1_to_ct2}{A ligand-receptor network list resulted from the function "get_expr_lr", signaling from ct1 to ct2}

\item{ligand_target_matrix_binary_ct2_to_ct1}{A matrix of 0s and 1s with target genes in rows and ligands in columns, indicates whether each target gene is identified as a target of each ligand, signaling to/in ct1}

\item{ligand_target_matrix_binary_ct1_to_ct2}{A matrix of 0s and 1s with target genes in rows and ligands in columns, indicates whether each target gene is identified as a target of each ligand, signaling to/in ct2}

\item{receptor_target_matrix_binary_ct2_to_ct1}{A matrix of 0s and 1s with target genes in rows and receptors in columns, indicates whether each target gene is identified as a target of each receptor, signaling to/in ct1}

\item{receptor_target_matrix_binary_ct1_to_ct2}{A matrix of 0s and 1s with target genes in rows and receptors in columns, indicates whether each target gene is identified as a target of each receptor, signaling to/in ct2}
}
\value{
myLRL: A list of two elements:
 myLRL$'L1->L2_L2->L1': LRloop network with L1->L2 and L2->L1 (L2 is a target of L1 and L1 is a target of L2)
 myLRL$'R1->L2_R2->L1': LRloop network with R1->L2 and R2->L1 (L2 is a target of R1 and L1 is a target of R2)
}
\description{
Get the LRloop L1_R1-L2_R2 pairs
}
