% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRLoopPartnerInfo.R
\name{getLRLoopPartnerInfo}
\alias{getLRLoopPartnerInfo}
\title{Get the number and list of LRLoop partners of LRpairs of interest.}
\usage{
getLRLoopPartnerInfo(LRpairs, from_to, LRLoop_network)
}
\arguments{
\item{LRpairs}{A matirx of the LRpairs of interes with two columns "from" and "to".}

\item{from_to}{"ct1_to_ct2" or "ct2_to_ct1" specifying if the LRpairs are from ct1 to ct2 or from ct2 to ct1.}

\item{LRLoop_network}{the LRLoop network between ct1 and ct2.}
}
\value{
LR_LRLoopPartnerInfo: A matrix with four columns "from", "to", "num_LRLoopPartner", "LRLoopPartner", in "from" and "to" are the LRpairs of interest, in "num_LRLoopPartner" and "LRLoopPartner" are the number and list of LRLoop partners of the LRpairs.  When "num_LRLoopPartner" is 0, the corresponding "LRLoopPartner" is "NA".
}
\description{
Get the number and list of LRLoop partners of LRpairs of interest.
}
