% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRscoreupdated_cutoff.R
\name{getLRscoreupdated_cutoff}
\alias{getLRscoreupdated_cutoff}
\title{Calculate an LRscores_LRL-cutoff based on the distribution of a specified column in LRscores_randLRL and create density plots of the specified score column of LRscores_LRL and LRscores_randLRL.}
\usage{
getLRscoreupdated_cutoff(p_value, LRscores_LRL, LRscores_randLRL, myscore_col)
}
\arguments{
\item{p_value}{Specifies the desired fraction of the scores in the column myscore_col in LRscores_randLRL that are above the cutoff.}

\item{LRscores_LRL}{The LRloops-updated LRscores.}

\item{LRscores_randLRL}{The random LRloops-updated LRscores.}

\item{myscore_col}{The name of the column in LRscores_LRL and LRscores_randLRL based on which the cutoff is calculated.}
}
\value{
Prints the cutoff value and creates density plots of the specified score column of LRscores_LRL and LRscores_randLRL.
}
\description{
Calculate an LRscores_LRL-cutoff based on the distribution of a specified column in LRscores_randLRL and create density plots of the specified score column of LRscores_LRL and LRscores_randLRL.
}
