% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnLRloops_rand.R
\name{getnLRloops_rand}
\alias{getnLRloops_rand}
\title{Construct a specified number of random LRloop networks between ct1 and ct2.}
\usage{
getnLRloops_rand(n, lr_ct1_to_ct2, lr_ct2_to_ct1, LRloops)
}
\arguments{
\item{n}{Number of random LRloop networks to be constructed.}

\item{lr_ct1_to_ct2}{A matrix with columns "from" and "to" of ligand-receptor pairs from ct1 to ct2.}

\item{lr_ct2_to_ct1}{A matrix with columns "from" and "to" of ligand-receptor pairs from ct2 to ct1.}

\item{LRloops}{The original LRloop network of all possible LRloops btween ct1 and ct2 of the LRpairs in lr_network.}
}
\value{
LRloops_rand: A list of n random LRloop networks.
}
\description{
Construct a specified number of random LRloop networks between ct1 and ct2.
}
