% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1R1L2R2Scores.R
\name{L1R1L2R2Score}
\alias{L1R1L2R2Score}
\title{Calculate the LRloop score of a L1-R1_L2-R2 pair}
\usage{
L1R1L2R2Score(L1, R1, L2, R2, m, LRscore, LoopScore)
}
\arguments{
\item{L1}{A number, the L1 value used for the calculation of the LRloop score}

\item{R1}{A number, the R1 value used for the calculation of the LRloop score}

\item{L2}{A number, the L2 value used for the calculation of the LRloop score}

\item{R2}{A number, the R2 value used for the calculation of the LRloop score}

\item{m}{A number, plays the role of a scalar in the calculation of LR scores when LRscore is set to 'scsigr' or individual_scale'}

\item{LRscore}{The method of calculating the LR scores, available options are 'mean', 'individual_scale', 'individual_scale_exp', 'product', 'bias_receptor' and 'scsigr'}

\item{LoopScore}{The method of calculating the LRloop score, available options are "ave_geo" and "ave_alg"}
}
\value{
The LRloop score of a L1-R1_L2-R2 pair
}
\description{
Calculate the LRloop score of a L1-R1_L2-R2 pair
}
