% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1R1L2R2ScoreHill.R
\name{L1R1L2R2ScoreHill}
\alias{L1R1L2R2ScoreHill}
\title{L1R1L2R2ScoreHill}
\usage{
L1R1L2R2ScoreHill(myscore, corrector, lambda, mu, k)
}
\arguments{
\item{myscore}{}

\item{corrector}{}

\item{lambda}{}

\item{mu}{}
}
\value{
z
}
\description{
L1R1L2R2ScoreHill
}
