% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRL_stats.R
\name{LRL_stats}
\alias{LRL_stats}
\title{Calculate some basic statistics on LRloops}
\usage{
LRL_stats(LRL_list)
}
\arguments{
\item{LRL_list}{List of identified LRloops calculated by the function "get_LRL"}
}
\description{
Calculate some basic statistics on LRloops
}
