% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRscores.R
\name{LRscore}
\alias{LRscore}
\title{Calculate the LRscore of an L-R pair}
\usage{
LRscore(L, R, m, method)
}
\arguments{
\item{L}{the ligand value used for the calculation of the LRscore}

\item{R}{the receptor value used for the calculation of the LRscore}

\item{m}{plays the role of a scalar in the calculation of the LRscore when the calculation method is set to 'scsigr' or individual_scale'}

\item{method}{The method of calculating the LRscore, available options are 'mean', 'individual_scale', 'individual_scale_exp', 'product', 'bias_receptor' and 'scsigr'. The method 'scsigr' was introduced in SingleCellSignalR.}
}
\value{
The LRscore of an L-R pair
}
\description{
Calculate the LRscore of an L-R pair
}
