% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_discrete_ligand_target_matrix_general.R
\name{make_discrete_ligand_target_matrix_general}
\alias{make_discrete_ligand_target_matrix_general}
\title{Identify the targets of each ligand's cognate receptors (1: TRUE; 0: FALSE)}
\usage{
make_discrete_ligand_target_matrix_general(
  ligands,
  receptor_target_matrix_binary,
  lr_network
)
}
\arguments{
\item{ligands}{Vector of ligand gene symbols of interest}

\item{receptor_target_matrix_binary}{Matrix of 0s and 1s with targets in rows and receptors in columns indicating whether each target gene is identified as a target of each receptor.}

\item{lr_network}{Ligand-receptor network matrix with columns "from" and "to".}
}
\value{
ligand_target_matrix_general_binary: Matrix of 0s and 1s, indicates whether each gene (in rows) is a target of at least one of the ligand's (in columns) cognate receptors
}
\description{
Identify the targets of each ligand's cognate receptors (1: TRUE; 0: FALSE)
}
