% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LRL_list.R
\name{make_LRL_list}
\alias{make_LRL_list}
\title{Find the regulatory relationships of all the (L2-R2)-(L1-R1) pairs of interest}
\usage{
make_LRL_list(
  L2R2_network,
  L1R1_network,
  ligand_target_matrix_binary_ct1,
  ligand_target_matrix_binary_ct2,
  receptor_target_matrix_binary_ct1,
  receptor_target_matrix_binary_ct2
)
}
\arguments{
\item{L2R2_network}{Ligand-receptor network matrix with columns "from" and "to", signaling from ct2 to ct1}

\item{L1R1_network}{Ligand-receptor network matrix with columns "from" and "to", signaling from ct1 to ct2}

\item{ligand_target_matrix_binary_ct1}{A matrix of 0s and 1s with target genes in rows and ligands in columns, indicates whether each target gene is identified as a target of each ligand, signaling to/in ct1}

\item{ligand_target_matrix_binary_ct2}{A matrix of 0s and 1s with target genes in rows and ligands in columns, indicates whether each target gene is identified as a target of each ligand, signaling to/in ct2}

\item{receptor_target_matrix_binary_ct1}{A matrix of 0s and 1s with target genes in rows and receptors in columns, indicates whether each target gene is identified as a target of each receptor, signaling to/in ct1}

\item{receptor_target_matrix_binary_ct2}{A matrix of 0s and 1s with target genes in rows and receptors in columns, indicates whether each target gene is identified as a target of each receptor, signaling to/in ct2}
}
\value{
p_A_a_list: A list of four elements:
 p_A_a[[all_reg_info]]: (L2-R2)-(L1-R1) pairs and the regulatory relationships between these genes (1: TRUE; 0: FALSE).
 p_A_a[[L1->L2 & L2->L1]]: (L2-R2)-(L1-R1) pairs that regulate each other, identified based on ligand_target_matrix_binary_ct1 and ligand_target_matrix_binary_ct2  
 p_A_a[[R1->L2 & R2->L1]]: (L2-R2)-(L1-R1) pairs that regulate each other, identified based on receptor_target_matrix_binary_ct1 and receptor_target_matrix_binary_ct2  
 p_A_a[[L1->L2 & L2->L1 & R1->L2 & R2->L1]]: (L2-R2)-(L1-R1) pairs that regulate each other, identified based on [ligand_target_matrix_binary_ct1 & receptor_target_matrix_binary_ct1] and [ligand_target_matrix_binary_ct2 & receptor_target_matrix_binary_ct2]
}
\description{
Find the regulatory relationships of all the (L2-R2)-(L1-R1) pairs of interest
}
