% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MostBasicLRinfo.R
\name{MostBasicLRinfo}
\alias{MostBasicLRinfo}
\title{Print: 
Number of L1-R1 pairs
Number of L1
Number of R1
Number of L2-R2 pairs
List the ligands L1
List the receptors R1
List the ligands L2
List the receptors R2}
\usage{
MostBasicLRinfo(LRloop_info)
}
\arguments{
\item{LRloop_info}{The list of the LRloop_network info calculated by function "LRL_info_collection"}
}
\description{
Print: 
Number of L1-R1 pairs
Number of L1
Number of R1
Number of L2-R2 pairs
List the ligands L1
List the receptors R1
List the ligands L2
List the receptors R2
}
