% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareBasics.R
\name{PrepareBasics}
\alias{PrepareBasics}
\title{Get a list of some basic data info}
\usage{
PrepareBasics(
  ct1obj,
  ct2obj,
  min_pct,
  geneset_ct1,
  geneset_ct2,
  lr_network,
  ligand_target_matrix_ct1_to_ct2,
  receptor_target_matrix_ct1_to_ct2,
  ligand_target_matrix_ct2_to_ct1,
  receptor_target_matrix_ct2_to_ct1,
  discrete_error_rate,
  discrete_cutoff_method,
  discrete_fdr_method
)
}
\arguments{
\item{ct1obj}{Seurat objects of cell type1 and cell type2}

\item{ct2obj}{Seurat objects of cell type1 and cell type2}

\item{min_pct}{Cutoff value on the detection rate of genes in each cell type, each condition for the identification of "expressed" genes. A number between 0 and 1. For example, min_pct = 0.1 means that for each cell type and condition, the genes that are detected in at least 10 percent of the cells are considered "expressed".}

\item{geneset_ct1}{Vectors of gene symbols considered as the target gene set of interest (for example, the differentailly expressed genes) in cell type1/cell type2 for the calculation of the nichenetr based ligand/receptor activity scores.}

\item{geneset_ct2}{Vectors of gene symbols considered as the target gene set of interest (for example, the differentailly expressed genes) in cell type1/cell type2 for the calculation of the nichenetr based ligand/receptor activity scores.}

\item{lr_network:}{Ligand-receptor network matrix with two columns "from" and "to", consisting of the gene symbols of the candidate ligands and receptors, respectively, in pairs.}

\item{ligand_target_matrix_ct1_to_ct2:}{Matrix of ligand to target regulatory potential scores calculated by nichenetr's algorithm, from ligands expressed in cell type1 to targets in cell type2.  Ligand gene symbols in columns and target gene symbols in rows.}

\item{ligand_target_matrix_ct2_to_ct1:}{Matrix of ligand to target regulatory potential scores calculated by nichenetr's algorithm, from ligands expressed in cell type2 to targets in cell type1.  Ligand gene symbols in columns and target gene symbols in rows.}

\item{receptor_target_matrix_ct1_to_ct2:}{Matrix of receptor to target regulatory potential scores calculated by nichenetr's algorithm, from receptors expressed in cell type2 (that are cognate receptors of ligands expressed in cell type1) to targets in cell type2.  Receptor gene symbols in columns and target gene symbols in rows.}

\item{receptor_target_matrix_ct2_to_ct1:}{Matrix of receptor to target regulatory potential scores calculated by nichenetr's algorithm, from receptors expressed in cell type1 (that are cognate receptors of ligands expressed in cell type2) to targets in cell type1.  Receptor gene symbols in columns and target gene symbols in rows.}

\item{discrete_error_rate:}{Value of the error_rate variable of the nichenetr package function "make_discrete_ligand_target_matrix" -- FDR for cutoff_method "fdrtool" and "distribution"; number between 0 and 1 indicating which top fraction of target genes should be returned for cutoff_method "quantile".}

\item{discrete_cutoff_method:}{Value of the cutoff_method variable of the nichenetr package function "make_discrete_ligand_target_matrix" -- Method to determine which genes can be considered as a target and which genes not, based on the regulatory probability scores. Possible options: "distribution", "fdrtool" and "quantile".}

\item{discrete_fdr_method:}{Value of the fdr_method variable of the nichenetr package function "make_discrete_ligand_target_matrix" -- Only relevant when cutoff_method is "fdrtool". Possible options: "global" and "local"}
}
\value{
Basics: A list with the following elements:
 Basics$ave_expr_ct1: A matrix of average gene expression values (LogNormalized) in each condition in cell type1.  Gene symbols in rows, conditions in columns
 Basics$ave_expr_ct2: A matrix of average gene expression values (LogNormalized) in each condition in cell type2.  Gene symbols in rows, conditions in columns
 Basics$pct_expr_ct1: A matrix of gene detection rates in each condition in cell type1.  Gene symbols in rows, conditions in columns
 Basics$pct_expr_ct2: A matrix of gene detection rates in each condition in cell type2.  Gene symbols in rows, conditions in columns
 Basics$thresh_expr_ct1: A binary matrix of 0s and 1s indicating if the detection rate of each gene passed the specified cutoff value min_pct (1: Yes; 0: No) in each condition in cell type1.  Gene symbols in rows, conditions in columns
 Basics$thresh_expr_ct2: A binary matrix of 0s and 1s indicating if the detection rate of each gene passed the specified cutoff value min_pct (1: Yes; 0: No) in each condition in cell type2.  Gene symbols in rows, conditions in columns
 Basics$genes_thresh_expr_ct1: A vector of gene symbols that are "expressed" (passed detection rate cutoff min_pct) in at least one condition in cell type1
 Basics$genes_thresh_expr_ct2: A vector of gene symbols that are "expressed" (passed detection rate cutoff min_pct) in at least one condition in cell type2
 Basics$lr_expr_ct1_to_ct2: A list with two elements:
 Basics$lr_expr_ct1_to_ct2$eachcondition: A list of ligand (expressed in cell type1)-receptor (expressed in cell type2) pairs "expressed" in each condition
 Basics$lr_expr_ct1_to_ct2$bind: Ligand (expressed in cell type1)-receptor (expressed in cell type2) pairs "expressed" in at least one condition
 Basics$lr_expr_ct2_to_ct1: A list with two elements:
 Basics$lr_expr_ct2_to_ct1$eachcondition: A list of ligand (expressed in cell type2)-receptor (expressed in cell type1) pairs "expressed" in each condition
 Basics$lr_expr_ct2_to_ct1$bind: Ligand (expressed in cell type2)-receptor (expressed in cell type1) pairs "expressed" in at least one condition
 Basics$ligand_activities_matrix_ct1_to_ct2: Matrix of regulatory potential scores of candidate ligands from cell type1 (calcuated by nichenetr's algorithm)
 Basics$ligand_activities_matrix_ct2_to_ct1: Matrix of regulatory potential scores of candidate ligands from cell type2 (calcuated by nichenetr's algorithm)
 Basics$receptor_activities_matrix_ct1_to_ct2: Matrix of regulatory potential scores of candidate receptors in cell type2 (that are cognate receptors of candidate ligands from cell type1) (calcuated by nichenetr's algorithm)
 Basics$receptor_activities_matrix_ct2_to_ct1: Matrix of regulatory potential scores of candidate receptors in cell type1 (that are cognate receptors of candidate ligands from cell type2) (calcuated by nichenetr's algorithm)
 Basics$myLRL: List of identified LRLoops:
 Basics$myLRL$R1->L2_R2->L1: Matrix of LRLoops with columns "L1", "R1", "L2" and "R2" where L1-R1 are candidate ligand-receptor pairs from cell type1 to cell type2, L2-R2 are candidate ligand-receptor pairs from cell type2 to cell type1; L2 is a top target of R1, L1 is a top target of R2
 Basics$myLRL$L1->L2_L2->L1: Matrix of LRLoops with columns "L1", "R1", "L2" and "R2" where L1-R1 are candidate ligand-receptor pairs from cell type1 to cell type2, L2-R2 are candidate ligand-receptor pairs from cell type2 to cell type1; L2 is a top target of L1, L1 is a top target of L2
}
\description{
Get a list of some basic data info
}
