% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_LRL_Score_options.R
\name{print_LRL_Score_options}
\alias{print_LRL_Score_options}
\title{Print available options for the function "plotLRL" variable "WhichL1R1L2R2Score"}
\usage{
print_LRL_Score_options(LRloop_info)
}
\arguments{
\item{LRloop_info}{The list of the LRloop_network info calculated by function "LRL_info_collection"}
}
\description{
Print available options for the function "plotLRL" variable "WhichL1R1L2R2Score"
}
