% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLRscores.R
\name{updateLRscores}
\alias{updateLRscores}
\title{Modify the LRscores of the LRs expressed from ct1 to ct2 (from ct2 to ct1) by the LRscores of their LRloop paired LRs from ct2 to ct1 (from ct1 to ct2).}
\usage{
updateLRscores(LRS, LRC, LRloops, from_to, LRCform = "max", lambda, mu, k)
}
\arguments{
\item{LRS}{The LRscores to be updated.  Its first two columns 'from' and 'to' are the LRpairs of interest.}

\item{LRC}{LRScore corrector matrix.  Its first two columns 'from' and 'to' are all the LRpairs in lr_network.}

\item{LRloops}{A matrix of all possible LRloops (in rows) between ct1 and ct2 of the LRpairs in lr_network.}

\item{from_to}{Specifies if the LRpairs in LRS are expressed from ct1 to ct2 or from ct2 to ct1.  Options are "ct1_to_ct2" or "ct2_to_ct1".}

\item{LRCform}{If an LRpair in LRS has multiple LRloop-partners in LRloops, it specifies how to get a single corrector score for it from the LRscores of these LRloop-partners in LRC. Options are "max", "mean" and "sum".}

\item{lambda}{Formula parameters in the function L1R1L2R2ScoreHill.}

\item{mu}{Formula parameters in the function L1R1L2R2ScoreHill.}

\item{k}{Formula parameters in the function L1R1L2R2ScoreHill.}
}
\value{
LRSC: LRscore matrix updated from LRS.  Compared to LRS, it has four more columns 'cond_max', 'cond_mean', 'twowaymax' and 'twowaymean'.
}
\description{
Modify the LRscores of the LRs expressed from ct1 to ct2 (from ct2 to ct1) by the LRscores of their LRloop paired LRs from ct2 to ct1 (from ct1 to ct2).
}
