% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLRL.R
\name{writeLRL}
\alias{writeLRL}
\title{Write the LRloop_info into .csv files}
\usage{
writeLRL(LRloop_info, filedir)
}
\arguments{
\item{LRloop_info}{The list of the LRloop_network info calculated by function "LRL_info_collection"}

\item{filedir}{Directory to save the files}
}
\description{
Write the LRloop_info into .csv files
}
