% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LRscores.R
\name{get_LRscores}
\alias{get_LRscores}
\title{Calculate the LRscore of each "expressed" ligand-receptor pair in each condition}
\usage{
get_LRscores(
  lr_expr,
  conditions,
  value.use_from,
  value.use_to,
  scalar,
  LRscore_method
)
}
\arguments{
\item{lr_expr}{A ligand-receptor network matrix with two columns "from" and "to".}

\item{conditions}{A vector of conditions of interest}

\item{value.use_from}{A numeric matrix with genes in rows and conditions in columns, selected as the ligand values to be used in the calculation of LRscores}

\item{value.use_to}{A numeric matrix with genes in rows and conditions in columns, selected as the receptor values to be used in the calculation of LRscores}

\item{scalar}{A number, plays the role of a scalar in the calculation of LRscores when the LRscore_method set to 'scsigr' or individual_scale'}

\item{LRscore_method}{The method of calculating the LRscores, available options are 'mean', 'individual_scale', 'individual_scale_exp', 'product', 'bias_receptor' and 'scsigr'}
}
\value{
A matrix with columns "ligand", "receptor" and the condition names that keeps record of the LRscore of each ligand-receptor pair in each condition
}
\description{
Calculate the LRscore of each "expressed" ligand-receptor pair in each condition
}
