% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1R1L2R2ScoreHill.R
\name{L1R1L2R2ScoreHill}
\alias{L1R1L2R2ScoreHill}
\title{L1R1L2R2ScoreHill}
\usage{
L1R1L2R2ScoreHill(myscore, corrector, lambda, mu, k)
}
\arguments{
\item{myscore}{myscore}

\item{corrector}{corrector}

\item{lambda}{lambda}

\item{mu}{mu}

\item{k}{k}
}
\value{
z z
}
\description{
L1R1L2R2ScoreHill
}
