% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRscores.R
\name{LRscore}
\alias{LRscore}
\title{Calculate the LRscore of an L-R pair}
\arguments{
\item{L}{A number, the ligand value used for the calculation of the LRscore}

\item{R}{A number, the receptor value used for the calculation of the LRscore}

\item{m}{A number, plays the role of a scalar in the calculation of the LRscore when the calculation method is set to 'scsigr' or individual_scale'}

\item{method}{The method of calculating the LRscore, available options are 'mean', 'individual_scale', 'individual_scale_exp', 'product', 'bias_receptor' and 'scsigr'. The method 'scsigr' was introduced in SingleCellSignalR. Reference: Simon Cabello-Aguilar, Mlissa Alame, Fabien Kon-Sun-Tack, Caroline Fau, Matthieu Lacroix, Jacques Colinge. Nucleic Acids Research, Volume 48, Issue 10, 04 June 2020, Page e55, https://doi.org/10.1093/nar/gkaa183}
}
\value{
The LRscore of an L-R pair
}
\description{
Calculate the LRscore of an L-R pair
}
