% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_LR_expr_score.R
\name{take_LR_expr_score}
\alias{take_LR_expr_score}
\title{Create a dataframe for the expr_scores of L1R1 or L2R2 pairs across all conditions stored in the list resulted from the function 'LRL_info_collection'}
\usage{
take_LR_expr_score(LRloop_info, LRpair)
}
\arguments{
\item{LRloop_info:}{A list resulted from the function 'LRL_info_collection'}

\item{LRpair:}{'L1R1' or "L2R2'}
}
\description{
Create a dataframe for the expr_scores of L1R1 or L2R2 pairs across all conditions stored in the list resulted from the function 'LRL_info_collection'
}
